Project 9 - Inventory App

+++ Project9/app/src/main/java/study/pmoreira/project9/utils/ImageUtils.java
@@ -38,11 +38,11 @@ public final class ImageUtils {
 
     public static Bitmap scale(Bitmap realImage, float maxImageSize, boolean filter) {
         float ratio = Math.min(
-                (float) maxImageSize / realImage.getWidth(),
-                (float) maxImageSize / realImage.getHeight());
+                maxImageSize / realImage.getWidth(),
+                maxImageSize / realImage.getHeight());
 
-        int width = Math.round((float) ratio * realImage.getWidth());
-        int height = Math.round((float) ratio * realImage.getHeight());
+        int width = Math.round(ratio * realImage.getWidth());
+        int height = Math.round(ratio * realImage.getHeight());
 
         return Bitmap.createScaledBitmap(realImage, width, height, filter);
     }
